Tabbify registration interface w/ Special Reg
