+++ client/src/pages/EventAdminRegistrations/SpecialRegTab.tsx
@@ -1,0 +1,22 @@
+import React from 'react';
+import {
+  InputGroup,
+  FormControl,
+  Container,
+  Row,
+  Col,
+  Tabs, Tab,
+} from 'react-bootstrap';
+
+function SpecialRegTab(props: EventAdminPageProps) {
+
+  return (
+    <Container>
+      <Row>
+        Hi
+      </Row>
+    </Container>
+  );
+}
+
+export default SpecialRegTab;
